update battery level function (untested still)

+++ include/Z2.hpp
@@ -148,7 +148,7 @@ public:
     ReportRate getReportRate();
 
     int getBatteryPercentage();
-    short getBatteryCharge();
+    int getBatteryCharge();
 };
